+++ .eslintrc.js
@@ -15,6 +15,8 @@ module.exports = {
     'plugin:@typescript-eslint/recommended',
     'prettier',
     'plugin:import/typescript',
+    'plugin:react/recommended',
+    'plugin:react-hooks/recommended',
   ],
   rules: {
     '@typescript-eslint/no-explicit-any': 'off',
@@ -39,5 +41,7 @@ module.exports = {
         ],
       },
     ],
+    'react/prop-types': 'off',
+    'react/react-in-jsx-scope': 'off',
   },
 }