+++ packages/blocks/lists/sortable-task-list/component.config.ts
@@ -8,12 +8,7 @@ export default {
   category: 'Application',
   subcategory: 'Lists',
   type: 'registry:block',
-  dependencies: [
-    '@chakra-ui/react',
-    '@chakra-ui/utils',
-    '@saas-ui/react',
-    'react-icons',
-  ],
+  dependencies: ['@chakra-ui/react', '@saas-ui/react', 'react-icons'],
   files: [
     {
       path: 'blocks/lists/sortable-task-list/sortable-task-list.tsx',