+++ apps/desktop/CHANGELOG.md
@@ -1,5 +1,32 @@
 # electron-app
 
+## 0.19.19
+
+### Patch Changes
+
+- Updated dependencies [045e219]
+  - @saas-ui-pro/feature-flags@0.18.0
+  - @saas-ui-pro/onboarding@0.16.0
+  - @saas-ui-pro/billing@0.14.0
+  - @saas-ui-pro/react@0.42.0
+
+## 0.19.18
+
+### Patch Changes
+
+- Updated dependencies [845e25a]
+  - @saas-ui-pro/react@0.41.0
+
+## 0.19.17
+
+### Patch Changes
+
+- Updated dependencies [d871bd0]
+- Updated dependencies [d871bd0]
+- Updated dependencies [ea1d123]
+  - @saas-ui-pro/react@0.40.0
+  - @saas-ui-pro/feature-flags@0.17.0
+
 ## 0.19.16
 
 ### Patch Changes