+++ packages/feature-flags/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @saas-ui/features
 
+## 0.18.0
+
+### Minor Changes
+
+- 045e219: Removed source files from published files
+
+## 0.17.0
+
+### Minor Changes
+
+- ea1d123: Fix SSR support of feature flag store, no longer exporting featureStore singleton.
+
 ## 0.16.0
 
 ### Minor Changes