+++ packages/kanban-core/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @saas-ui-pro/kanban-core
 
+## 0.5.0
+
+### Minor Changes
+
+- 045e219: Removed source files from published files
+
+## 0.4.1
+
+### Patch Changes
+
+- 88ecff5: Fixed issue where onItemsChange was passed to dom element
+
+## 0.4.0
+
+### Minor Changes
+
+- 2f2c002: Moving between columns can now be prevented using the isMoveAllowed handler
+- 2f2c002: Renamed onChange to onItemsChange
+
 ## 0.3.2
 
 ### Patch Changes